Add failing test for viewportHistogram (+ test refactor)
